# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation APIRest for Esteban Project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/videos": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "videos list",
                "description": "Get videos.",
                "operationId": "getAllVideos",
                "responses": {
                    "200": {
                        "description": "get all videos",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Videos"
                ],
                "summary": "create a new video",
                "operationId": "newVideo",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "video"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Mi moto nueva"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "-------"
                                    },
                                    "video": {
                                        "description": "video",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "response create new video",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Los datos proporcionados no son válidos",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/videos/byUser/{user_id}": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "video",
                "description": "Get videos by user Id.",
                "operationId": "getVideosByUserId",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": "3"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get videos by user id",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/videos/{id}": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "video",
                "description": "Get video Id.",
                "operationId": "getVideo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of video",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": "3"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get video",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Videos"
                ],
                "summary": "delete videos",
                "description": "destroy video.",
                "operationId": "DeleteVideo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": "3"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete video",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/videos/update/{id}": {
            "post": {
                "tags": [
                    "Videos"
                ],
                "summary": "update a video",
                "operationId": "updateVideo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": "3"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Mi moto vieja"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Vendida hace poco"
                                    },
                                    "video": {
                                        "description": "video",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "response update video",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}